+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Changelog
 
 ## [3.2.0](https://github.com/maxgaurav/indexeddb-orm/releases/tag/v3.1.1) _(Latest Release)_
+### Bug Fixes
+- Fix **openTransaction** on model not returning ORM Class.
 
 ### Added
 - Sync actions **(sync, syncIndex, syncIndexAll)** to update and retrieve the updated record on success.
 - Added schema configuration to allow opt-in for maintaining synced on date when calling sync actions. Will add/update **syncOn** attribute
 - Added **saveIndex** to save data on index. It's optimized for single action transaction over update
 - Added **saveAllIndex** to save data on all matching index.
+
 
 ### Other
 - Refactored **model.ts** file to multiple files to maintain single file single responsibility principle.